[Refactor] Clean up duplicated code in matmul_op/batch_matmul_op and associated tests.
